Triple backticks indicate MIT license in all files
Just clarifying the terms. Originally, all my code was in README.md --
which explains the previous wording. What I actually meant is that all
code snippets are under the MIT license (or, rather, *also* under the
MIT license, because the code snippets are really dual licensed: MIT and
CC BY-NC-SA 4.0 -- but this was hopefully clear since the beginning.)